+++ examples/clean_architecture/android/app/src/main/kotlin/com/example/clean_architecture/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.clean_architecture
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity()