+++ examples/clean_architecture/lib/domain/usecase/add_todo.dart
@@ -1,0 +1,25 @@
+import '../../data/repository/todos_impl.dart';
+import '../model/todo.dart';
+
+class AddTodo {
+  final TodosRepository _todosRepository;
+
+  AddTodo(this._todosRepository);
+
+  Future<Todo> call({
+    required String title,
+    bool completed = false,
+  }) async {
+    final id = await _todosRepository.add(
+      Todo.newTodo(
+        title: title,
+        completed: completed ? 1 : 0,
+      ),
+    );
+    final result = await _todosRepository.get(id);
+    if (result == null) {
+      throw Exception('Todo not found');
+    }
+    return result;
+  }
+}