+++ examples/clean_architecture/lib/domain/usecase/delete_completed_todos.dart
@@ -1,0 +1,11 @@
+import '../../data/repository/todos_impl.dart';
+
+class DeleteCompletedTodos {
+  final TodosRepository _todosRepository;
+
+  DeleteCompletedTodos(this._todosRepository);
+
+  Future<void> call() async {
+    await _todosRepository.removeCompleted();
+  }
+}