+++ examples/clean_architecture/lib/domain/usecase/delete_todo.dart
@@ -1,0 +1,15 @@
+import '../../data/repository/todos_impl.dart';
+
+class DeleteTodo {
+  final TodosRepository _todosRepository;
+
+  DeleteTodo(this._todosRepository);
+
+  Future<void> call(int id) async {
+    final current = await _todosRepository.get(id);
+    if (current == null) {
+      throw Exception('Todo not found');
+    }
+    await _todosRepository.remove(current);
+  }
+}