+++ examples/clean_architecture/lib/domain/usecase/get_todos.dart
@@ -1,0 +1,12 @@
+import '../../data/repository/todos_impl.dart';
+import '../model/todo.dart';
+
+class GetTodos {
+  final TodosRepository _todosRepository;
+
+  GetTodos(this._todosRepository);
+
+  Stream<List<Todo>> call() {
+    return _todosRepository.watchAll();
+  }
+}