+++ examples/clean_architecture/lib/domain/usecase/update_todo.dart
@@ -1,0 +1,12 @@
+import '../../data/repository/todos_impl.dart';
+import '../model/todo.dart';
+
+class UpdateTodo {
+  final TodosRepository _todosRepository;
+
+  UpdateTodo(this._todosRepository);
+
+  Future<void> call(Todo todo) async {
+    await _todosRepository.update(todo);
+  }
+}